+++ frontend/src/components/pages/listings/index/ListingItem.tsx
@@ -17,7 +17,7 @@ import timezone from "dayjs/plugin/timezone";
 import utc from "dayjs/plugin/utc";
 import Link from "next/link";
 import ReactMarkdown from "react-markdown";
-import * as markdownComponents from "@components/pages/listings/markdown/components";
+import * as markdownComponents from "@components/markdown/components";
 
 dayjs.extend(timezone);
 dayjs.extend(utc);
